+++ main/modules/lvgl/indev_i2c.h
@@ -7,7 +7,7 @@
 
 
 bool indev_nav_read_i2c(indev_driver_spec_t* spec, uint8_t * byte) ;
-uint8_t indev_nav_take_value(indev_driver_spec_t * spec) ;
+uint32_t indev_nav_take_value(indev_driver_spec_t * spec) ;
 
 void be_indev_i2c_init() ;
 void be_indev_i2c_require(JSContext *ctx, JSValue pkg, JSValue baseProto) ;
